requirements.txt
aiohttp==3.7.4.post0
anyio==3.3.4
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.7
dataclasses==0.6
deprecation==2.1.0
discord==1.7.3
discord.py==1.7.3
gotrue==0.2.0
h11==0.12.0
httpcore==0.13.7
httpx==0.19.0
idna==2.10
iniconfig==1.1.1
multidict==5.2.0
packaging==21.2
pluggy==1.0.0
postgrest-py==0.5.0
py==1.11.0
pyparsing==2.4.7
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.19.1
realtime-py==0.1.3
requests==2.25.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
supabase==0.0.3
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.7
websockets==9.1
yarl==1.7.2
